Removes unnecessary dependencies
Currently the env file is an export of 'base' which has far more
dependencies that we need for this environment. Instead this PR only
includes dependencies needed to run this code.

As a result, you will need to create a new environment (e.g. don't
attempt to update base) and you can do that with

```
conda env create -f environment.yml
```

which will create an environment called `ssheet-analyser` - if you want
to use a different name, append `-n NAME`.